Move a few common methods into a BaseAdapter
typeOf, compare, equals and callFunction are all generic enough to be shared among all implementations. JacksonAdapter can implement typeOf with a switch statement so it still has its own implementation of that method.
