+++ jmespath-jackson/src/main/java/io/burt/jmespath/jackson/JacksonAdapter.java
@@ -14,20 +14,19 @@ import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
-import io.burt.jmespath.Adapter;
+import io.burt.jmespath.BaseAdapter;
 import io.burt.jmespath.JmesPathType;
 import io.burt.jmespath.function.FunctionRegistry;
 import io.burt.jmespath.function.ExpressionOrValue;
 
 import static com.fasterxml.jackson.databind.node.JsonNodeType.*;
 
-public class JacksonAdapter implements Adapter<JsonNode> {
+public class JacksonAdapter extends BaseAdapter<JsonNode> {
   private final ObjectMapper jsonParser;
-  private final FunctionRegistry functionRegistry;
 
   public JacksonAdapter() {
+    super();
     this.jsonParser = new ObjectMapper();
-    this.functionRegistry = FunctionRegistry.createDefaultRegistry();
   }
 
   @Override
@@ -142,30 +141,6 @@ public class JacksonAdapter implements Adapter<JsonNode> {
   }
 
   @Override
-  public int compare(JsonNode value1, JsonNode value2) {
-    if (value1.getNodeType() == value2.getNodeType()) {
-      switch (value1.getNodeType()) {
-        case ARRAY:
-        case BINARY:
-        case OBJECT:
-        case BOOLEAN:
-        case MISSING:
-        case NULL:
-        case POJO:
-          return value1.equals(value2) ? 0 : -1;
-        case STRING:
-          return value1.textValue().compareTo(value2.textValue());
-        case NUMBER:
-          return Double.compare(value1.doubleValue(), value2.doubleValue());
-        default:
-          throw new IllegalStateException(String.format("Unknown node type encountered: %s", value1.getNodeType()));
-      }
-    } else {
-      return -1;
-    }
-  }
-
-  @Override
   public JsonNode getProperty(JsonNode value, String name) {
     return nodeOrNullNode(value.get(name));
   }
@@ -230,21 +205,11 @@ public class JacksonAdapter implements Adapter<JsonNode> {
     return JsonNodeFactory.instance.numberNode(n);
   }
 
-  @Override
-  public JsonNode callFunction(String name, List<ExpressionOrValue<JsonNode>> arguments) {
-    return functionRegistry.callFunction(this, name, arguments);
-  }
-
   private JsonNode nodeOrNullNode(JsonNode node) {
     if (node == null) {
       return JsonNodeFactory.instance.nullNode();
     } else {
       return node;
     }
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    return o instanceof JacksonAdapter;
   }
 }
